requirements.txt
Django==2.2.4
djangorestframework==3.10.2
djangorestframework-simplejwt==4.3.0
PyJWT==1.7.1
pytz==2021.3
sqlparse==0.4.2
